cells outside grid are now considered as alive, (change of type for parameters on count neighbours) size_t (unsigned) -->	long long int (signed)

+++ Grid.hpp
@@ -40,7 +40,7 @@ public:
 	bool isAlive(const sf::Vector2i& position) const;
 	bool isAlive(const long long int &x, const long long int &y)const ;
 	
-	size_t countNeighbours(const size_t &x, const size_t & y) const ;
+	size_t countNeighbours(const long long int &x, const long long int & y) const ;
 	size_t countNeighbours(const sf::Vector2i & position) const;
 	
 	void computeNextIteration();
